Dependabot::GoModules::ReplaceStubber
